Created IronPython WCF Client (Hooray!).
Although when I simply used the IronPython console, it kept "hanging."
(Saw the spinning ball forever. Did not even respond to Stop Execution.)

I then ran everything in a terminal and it all worked. Apparently simply
invoking the remote service by `proxy.HelloIndigo()` and **not**
printing the result put the console in an odd state. By adding invoking
the service as part of a print statement, `print proxy.HelloIndigo()`,
it worked like a champ.

Additionally, terminating the console hosting the IronPython client
terminated the server. I do not know why (and did not try to
investigate).