Create skeleton (console) host project.
This skeleton is an empty console project that simply waits for the user
to press the <ENTER> key and then terminates.

Later, I will add to the code to create an actual service host.

Because it is a skeleton, I can actually copy the `Host` project to
create a similar `Client` project.