Create client service reference. Client works!
Add a service reference to the `Client` project following the
instructions in the book. Build all projects. Set startup project to
include both `Host` and `Client` if not already done. Start all
projects.

The one configuration detail that I noticed was that the
`bindingConfiguration` attribute was "basicHttp" - **not** an empty
string. I'll go back and change that in the `unit_test_broke` branch to
see if the unit tests now pass.